# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'QMS API Documentation'
  description: 'Secure messenger API with end-to-end encryption'
  contact:
    name: Heruvim6w
    email: heruvim.6w@gmail.com
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'QMS API Server'
  -
    url: 'http://localhost:8000'
    description: 'Development Server'
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
paths:
...
  /api/v1/register:
    post:
      tags:
        - Authentication
      summary: 'Register a new user account'
      description: 'Create a new user account with email and password. User automatically receives a unique UIN (8-digit identifier) and can optionally set a custom username later. RSA key pair is generated for end-to-end encryption.'
      operationId: register
      requestBody:
        description: 'User registration data'
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - email
                - password
                - password_confirmation
              properties:
                name:
                  description: 'User full name'
                  type: string
                  maxLength: 255
                  minLength: 2
                  example: 'John Doe'
                email:
                  description: 'Valid email address (must be unique)'
                  type: string
                  format: email
                  example: john@example.com
                password:
                  description: 'Password (min 8 chars, must include uppercase, lowercase, digit)'
                  type: string
                  format: password
                  minLength: 8
                  example: SecurePass123!
                password_confirmation:
                  description: 'Password confirmation (must match password)'
                  type: string
                  format: password
                  example: SecurePass123!
              type: object
      responses:
        201:
          description: 'User registered successfully. JWT token issued for immediate access.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        422:
          description: 'Validation error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        500:
          description: 'Server error (e.g., encryption key generation failed)'
  /api/v1/login:
    post:
      tags:
        - Authentication
      summary: 'Authenticate user - Step 1 (Initiate login)'
      description: 'Start the login process. For new devices, sends a confirmation email with a link valid for 3 hours. For known devices, returns JWT token immediately. Users can login with either email or UIN.'
      operationId: login
      requestBody:
        description: 'Login credentials'
        required: true
        content:
          application/json:
            schema:
              required:
                - login
                - password
              properties:
                login:
                  description: 'Email address or UIN (8 digits)'
                  type: string
                  example: john@example.com
                password:
                  description: 'User password'
                  type: string
                  format: password
                  example: SecurePass123!
                device_name:
                  description: 'Device name (optional, for tracking)'
                  type: string
                  example: 'iPhone 13'
                  nullable: true
              type: object
      responses:
        200:
          description: 'Check your email for confirmation link (new device) or token issued immediately (known device)'
          content:
            application/json:
              schema:
                oneOf:
                  -
                    $ref: '#/components/schemas/LoginStep1Response'
                  -
                    $ref: '#/components/schemas/LoginStep2Response'
        401:
          description: 'Invalid credentials (wrong email/UIN or password)'
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: 'Invalid credentials'
                type: object
        422:
          description: 'Validation error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /api/v1/login/confirm:
    post:
      tags:
        - Authentication
      summary: 'Confirm login - Step 2 (Verify email token)'
      description: 'Verify the confirmation token received via email. After successful verification, JWT token is issued. Token must be confirmed within 3 hours or it expires.'
      operationId: confirmLogin
      requestBody:
        description: 'Confirmation token from email'
        required: true
        content:
          application/json:
            schema:
              required:
                - token
              properties:
                token:
                  description: 'Token received in confirmation email'
                  type: string
                  example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...
              type: object
      responses:
        200:
          description: 'Login confirmed successfully. JWT token issued.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmLoginResponse'
        400:
          description: 'Token is invalid or expired (expired after 3 hours)'
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                    example: 'Invalid or expired token'
                type: object
  '/api/v1/login/confirm/{token}':
    get:
      tags:
        - Authentication
      summary: 'Confirm login via web link (browser)'
      description: 'Verification endpoint for web-based confirmation link from email. Can redirect to mobile app or return token directly. Primarily used for browser/web clients.'
      operationId: confirmLoginWeb
      parameters:
        -
          name: token
          in: path
          description: 'Confirmation token from email'
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 'Login confirmed. Redirect to mobile app or return token.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmLoginResponse'
        302:
          description: 'Redirect to mobile app with token parameter'
          headers:
            Location:
              description: 'Redirect URL with token parameter'
              schema:
                type: string
                format: uri
        400:
          description: 'Token is invalid or expired'
          content:
            application/json:
              schema:
                properties:
                  error:
                    type: string
                type: object
  /api/v1/logout:
    post:
      tags:
        - Authentication
      summary: 'Logout and invalidate JWT token'
      description: 'Invalidate the current JWT token and end the session. Existing sessions on other devices remain valid. Must be called before discarding the token.'
      operationId: logout
      responses:
        200:
          description: 'Successfully logged out'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Successfully logged out'
                type: object
        401:
          description: 'Unauthenticated (no valid token provided)'
      security:
        -
          bearerAuth: []
  /api/v1/refresh:
    post:
      tags:
        - Authentication
      summary: 'Refresh JWT token'
      description: 'Extend the current session by issuing a new JWT token. Call this when token is about to expire (check expires_in value). Old token becomes invalid immediately.'
      operationId: refresh
      responses:
        200:
          description: 'Token refreshed successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmLoginResponse'
        401:
          description: 'Token is invalid or expired and cannot be refreshed'
      security:
        -
          bearerAuth: []
  /api/v1/me:
    get:
      tags:
        - Authentication
      summary: 'Get the authenticated user profile'
      description: 'Retrieve the complete profile information of the authenticated user, including personal data, status, and settings.'
      operationId: me
      responses:
        200:
          description: 'Current user details'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        401:
          description: 'Unauthenticated (no valid token provided)'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    ValidationError:
      title: ValidationError
      description: 'Validation error response'
      properties:
        message:
          description: 'Error message'
          type: string
          example: 'The given data was invalid.'
        errors:
          description: 'Validation errors'
          type: object
          example:
            email:
              - 'The email field is required.'
            password:
              - 'The password field is required.'
      type: object
...
    RegisterResponse:
      title: 'Register Response'
      description: 'Successful user registration response'
      properties:
        user:
          properties:
            id:
              type: string
              format: uuid
              example: 550e8400-e29b-41d4-a716-446655440000
            name:
              type: string
              example: 'John Doe'
            email:
              type: string
              format: email
              example: john@example.com
            uin:
              description: '8-digit UIN generated at registration'
              type: string
              example: '12345678'
            username:
              type: string
              example: null
              nullable: true
          type: object
        access_token:
          description: 'JWT access token'
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          description: 'Token TTL in seconds'
          type: integer
          example: 3600
      type: object
    LoginStep1Response:
      title: 'Login Step 1 Response - New Device'
      description: 'Response when login is from a new device (requires email confirmation)'
      properties:
        message:
          type: string
          example: 'Confirmation email sent. Link is valid for 3 hours.'
        requires_confirmation:
          type: boolean
          example: true
      type: object
    LoginStep2Response:
      title: 'Login Step 2 Response - Known Device'
      description: 'Immediate login response for known devices'
      properties:
        access_token:
          description: 'JWT access token'
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          example: 3600
        user:
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            email:
              type: string
              format: email
            uin:
              type: string
            username:
              type: string
              nullable: true
          type: object
        requires_confirmation:
          type: boolean
          example: false
      type: object
    ConfirmLoginResponse:
      title: 'Confirm Login Response'
      description: 'Response after email confirmation'
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          example: 3600
      type: object
    UserProfileResponse:
      title: 'User Profile'
      description: 'Complete user profile information'
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        uin:
          description: '8-digit unique identifier'
          type: string
        username:
          description: 'Custom username (3-20 chars, latin only)'
          type: string
          nullable: true
        status:
          description: 'Current online status'
          type: string
          enum:
            - online
            - offline
        online_status:
          description: 'Selected status mood'
          type: string
          enum:
            - online
            - chatty
            - angry
            - depressed
            - home
            - work
            - eating
            - away
            - unavailable
            - busy
            - do_not_disturb
        custom_status:
          description: 'User custom status text with emoji support'
          type: string
          maxLength: 50
          nullable: true
        last_seen_at:
          description: 'Last online timestamp'
          type: string
          format: date-time
          nullable: true
        locale:
          description: 'User interface language'
          type: string
          example: ru
        created_at:
          type: string
          format: date-time
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  -
    bearerAuth: []